Add files via upload
HTML, CSS, JavaScript Practice

In this project, I utilized HTML, CSS, and JavaScript to create a dynamic and visually appealing bio. The HTML provided the structure of the bio, including headings, paragraphs, and lists. CSS was used to style the elements, defining their appearance, layout, and overall design to ensure a polished and professional presentation.

JavaScript played a crucial role in adding interactivity and functionality to the bio. Through JavaScript, I implemented features such as interactive buttons, hover effects, and animations, enhancing the user experience and making the bio more engaging.

Overall, this project served as an opportunity to strengthen my skills in front-end web development and demonstrate proficiency in HTML, CSS, and JavaScript. It showcases my ability to create responsive and interactive web content, essential skills for modern web development projects.